requirements.txt
alembic==1.4.2
appdirs==1.4.4
Babel==2.8.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dnspython==2.0.0
dominate==2.5.2
elasticsearch==7.9.1
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.1.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.3
flup==1.0.3
guess-language==0.2
guess-language-spirit==0.5.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.0.0
loginform==1.2.0
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
pbr==5.4.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
rq==1.5.1
six==1.15.0
SQLAlchemy==1.3.19
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
Tempita==0.5.2
urllib3==1.25.10
virtualenv==20.0.31
visitor==0.1.3
Werkzeug==1.0.1
Whoosh==2.7.4
WTForms==2.3.3
